refactor(neovim): replace custom mason utility with plugin (#2)
Replaced the custom utils.mason module with the mason-tool-installer.nvim plugin. This simplifies the installation and management of mason packages for formatters and linters.
The conform.lua and nvim-lint.lua configurations were updated to use the new plugin, and the plugins/conform.lua and plugins/nvim-lint.lua files were updated to reflect the new dependency. The utils/mason.lua
file was removed as it is no longer needed.